+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Features:
   - [language-server-code-transform] Add command to import class - @dantleech
   - [language-server-completion] Automatically import class on completion confirm - @dantleech
   - [code-transform] Consider current class as a potential conflict for imports - @dantleech
+  - [completion] Indexed class name and function completion - @dantleech
 
 Improvements: